+++ PROYECTO EMILY REP/Assets/Scripts/Character/Player/PlayerNetworkManager.cs
@@ -50,13 +50,13 @@ namespace KC
             }
         }
 
-        public void SetNewMaxHealthValue(float oldVitality, float newVitality)
+        public void SetNewMaxHealthValue(int oldVitality, int newVitality)
         {
             maxHealth.Value = player.playerStatsManager.CalculateHealthBasedOnVitalityLevel(newVitality);
             PlayerUIManager.instance.playerUIHudManager.SetMaxHealthValue(maxHealth.Value);
             currentHealth.Value = maxHealth.Value;
         }
-        public void SetNewMaxStaminaValue(float oldEndurance, float newEndurance)
+        public void SetNewMaxStaminaValue(int oldEndurance, int newEndurance)
         {
             maxStamina.Value = player.playerStatsManager.CalculateStaminaBasedOnEnduranceLevel(newEndurance);
             PlayerUIManager.instance.playerUIHudManager.SetMaxStaminaValue(maxStamina.Value);
